TypeScriptify `hasActiveOffer`
